requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
idna==3.11
joblib==1.5.3
mysqlclient==2.2.7
numpy==2.4.0
pandas==2.3.3
pillow==12.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
